28/12/2020

+++ Estructura_Codigo_funciones_resta_suma_producto.c
@@ -1,46 +1,46 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-int suma(int x , int y);
-int resta(int w, int z);
-int producto(int u, int v);
+float suma(float x , float y);
+float resta(float w, float z);
+float producto(float u, float v);
 
 int main(){
-    int x,y,r;
-    int w,z,r2;
-    int u,v,r3;
+    float x,y,r;
+    float w,z,r2;
+    float u,v,r3;
     printf("Ingrese el primer valor a sumar: \t");
-    scanf("%d",&x);
+    scanf("%f",&x);
     printf("Ingrese el segundo valor a sumar: \t");
-    scanf("%d",&y);
+    scanf("%f",&y);
     r=suma(x,y);
-    printf("El reultado de la suma es : \t %d", r);
+    printf("El reultado de la suma es : \t %f", r);
     printf("\n-------------\n");
     printf("Ingrese el primer valor que se restara: \t");
-    scanf("%d",&w);
+    scanf("%f",&w);
     printf("Ingrese el segundo valor que resta : \t");
-    scanf("%d",&z);
+    scanf("%f",&z);
     r2=resta(w,z);
-     printf("El reultado de la resta  es : \t %d", r2);
+     printf("El reultado de la resta  es : \t %f", r2);
       printf("\n-------------\n");
        printf("Ingrese el primer valor que se multiplicara: \t");
-    scanf("%d",&u);
+    scanf("%f",&u);
     printf("Ingrese el segundo valor que multiplicara : \t");
-    scanf("%d",&v);
+    scanf("%f",&v);
     r3=producto(u,v);
-    printf("El reultado del producto es : \t %d", r3);
+    printf("El reultado del producto es : \t %f", r3);
     return 0;
 }
-int suma (int x , int y){
+float suma (float x , float y){
     
     return x+y;
 }
-int resta(int w, int z){
+float resta(float w, float z){
     
     return w-z;
 }
 
-int producto(int u, int v){
+float producto(float u, float v){
     
     return u*v;
 }
